requirements.txt
pyyaml
autogen-core
streamlit
numpy
pandas
autogen-agentchat
autogen-ext[openai]
autogen-core
autogen-ext
typing_extensions
python-pptx
azure-search-documents
pillow
typing
pydantic
openai
pypdfium2
tiktoken
langchain